+++ src/components/Rightbar.jsx
@@ -1,22 +1,22 @@
 // components
-import {Container, Typography} from '@mui/material';
-import Home from '@mui/icons-material/Home';
+import {Container} from '@mui/material';
 import Explore from './Explore.jsx';
 
 import makeStyles from '@mui/styles/makeStyles';
+import React from 'react';
 
 const useStyles = makeStyles((theme) => ({
   container: {
-      padding: theme.spacing(5),
+      padding: 5,
   }
 }));
 
-function Rightbar({ host, username, password, data, keywords, updateData }) {
+function Rightbar({ data, keywords, updateData }) {
   const classes = useStyles();
 
   return (
     <Container className={classes.container}>
-        <Explore host={host} username={username} password={password} keywords={keywords} data={data} updateData={updateData}/>
+        <Explore keywords={keywords} data={data} updateData={updateData}/>
     </Container>
   );
 }
