+++ src/pages/LoginPage.jsx
@@ -2,15 +2,15 @@
 import React, {  } from 'react';
 
 // components
-import SignIn from '../components/SignIn.js'
+import SignIn from '../components/SignIn';
 
 import makeStyles from '@mui/styles/makeStyles';
 import { Grid } from '@mui/material';
+import SigninPresentation from 'components/SignInPresentation';
 
 const useStyles = makeStyles((theme) => ({
   container: {
     minHeight: "100vh",
-    paddingTop: 50,
     backgroundColor: "#F5F5F5",
   },
 }));
@@ -19,12 +19,12 @@ function LoginPage({username, updateUsername, password, updatePassword }) {
   const classes = useStyles();
 
   return (
-        <Grid container spacing={2} className={classes.container}>
+        <Grid container spacing={0} className={classes.container}>
           <Grid item xs={6}>
             <SignIn username={username} updateUsername={updateUsername} password={password} updatePassword={updatePassword} />
           </Grid>
-          <Grid item xs={6}>  
-            
+          <Grid item xs={6}>
+            <SigninPresentation/>
           </Grid>
         </Grid>
   );
